--- a/bindings/python/Cargo.toml
+++ b/bindings/python/Cargo.toml
@@ -12,7 +12,6 @@ crate-type = ["cdylib"]
 pyo3 = { version = "0.17.2", features = ["extension-module"] }
 memmap2 = "0.5"
 serde_json = "1.0"
-libloading = "0.7"
 
 [dependencies.safetensors]
 version = "*"