--- a/bindings/python/Cargo.toml
+++ b/bindings/python/Cargo.toml
@@ -9,7 +9,7 @@ name = "safetensors_rust"
 crate-type = ["cdylib"]
 
 [dependencies]
-pyo3 = { version = "0.17.2", features = ["extension-module"] }
+pyo3 = { version = "0.18.1", features = ["extension-module"] }
 memmap2 = "0.5"
 serde_json = "1.0"