Beats my expectations. I wonder what I did do to not see this perf
bottleneck/improvements.

# Main

```
------------------------------------------------ benchmark: 1 tests ------------------------------------------------
Name (time in ms)            Min       Max      Mean  StdDev    Median     IQR  Outliers     OPS  Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------
Titan RTX:                164.2941  178.6739  169.0371  6.0686  166.3407  8.5996       1;0  5.9159       5           1
T4 + Windows              614.6252  677.2240  648.6450  25.4562  648.0336  41.6149
3090                      339.0677  348.0977  341.3016  3.8854  339.2288  3.6862       1;0  2.9300       5           1
```

# This PR

```
------------------------------------------------ benchmark: 1 tests ------------------------------------------------
Name (time in ms)            Min       Max      Mean  StdDev    Median     IQR  Outliers     OPS  Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------
Titan RT                      93.3372  105.2049  96.7188  4.9857  94.4572  5.7376       1;0  10.3392       5           1
T4 + Windows                 374.2898  383.0617  377.1890  3.7953  375.3404  5.6582
3090                         339.2538  348.5638  341.6615  3.9584  339.6492  3.8666       1;0  2.9269       5           1
```